feat: Separate out component types
