+++ src/components/index.js
@@ -1,7 +1,8 @@
-export { default as Button } from './Button';
-export { default as Heading } from './Heading';
-export { default as Heading2 } from './Heading/Heading2';
-export { default as Heading3 } from './Heading/Heading3';
-export { default as Heading4 } from './Heading/Heading4';
-export { default as Nautilus } from './Nautilus';
-export { default as Paragraph } from './Paragraph';
+export { default as Nautilus } from './hoc/Nautilus';
+
+export { default as Button } from './ui/Button';
+export { default as Heading } from './ui/Heading';
+export { default as Heading2 } from './ui/Heading/Heading2';
+export { default as Heading3 } from './ui/Heading/Heading3';
+export { default as Heading4 } from './ui/Heading/Heading4';
+export { default as Paragraph } from './ui/Paragraph';
