+++ styleguide/docs/components/ui.md
@@ -1,0 +1,3 @@
+Use UI Components to display elements visually on a screen. Think of these as the "UI Kit" portion of Nautilus; if you want to display a header, an image, or tabular data—you should be using a UI component.
+
+Our UI Components are _very_ low-level and should be used as replacement for HTML elements. This is to provide a more consistent API around content (that adheres to the rules of our design system) and to provide a platform-agnostic UI library—you should be able to use Nautilus components in both a React and React Native app.
